requirements.txt
Flask>=3.0
launchdarkly-server-sdk>=9.0
python-dotenv>=1.0
pytest>=8.0
requests>=2.31
